Implement support for cilium-health in ENI mode
[ upstream commit 26308b6304ec9bd1326ac1cdb715ab51cb5b198a ]

In ENI mode, rules and routes are created for each endpoint (and
therefore pod) in order to direct traffic to the appropriate ENI device.
This logic was only running in the CNI plugin. Since cilium-health is
not a pod, its rules and routes needed for traffic to flow were missing.

This commit implements creating the necessary rules and routes for
cilium-health in ENI mode. This commit also unifies the installation
logic of the rules and routes in `pkg/aws/eni/routing`, allowing the CNI
plugin and the cilium-health creation code to share a common
implementation.

Fixes #10810

Signed-off-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>