k8s: Require use of discovery API to be enabled
[ upstream commit 6fcf0cdd667945e59c764242a00e923c8cf3bc77 ]

The discovery request consists of several API calls to the apiserver. It
has already been observed that the k8s client can stsrt throttling
during startup phase.

Disable the use of the discovery API by default and rely on individual
API probing but allow users to benefit from it by opting into it.

Signed-off-by: Thomas Graf <thomas@cilium.io>
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>